Upgraded openexchangerates to v3 framework
