+++ .changeset/yellow-knives-hear.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/openexchangerates-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
