+++ packages/core/legos/tsconfig.json
@@ -359,9 +359,6 @@
       "path": "../../sources/openexchangerates"
     },
     {
-      "path": "../../sources/openexchangerates-test"
-    },
-    {
       "path": "../../sources/orchid-bandwidth"
     },
     {
