--- packages/sources/openexchangerates-test/CHANGELOG.md
@@ -1,46 +1,0 @@
-# @chainlink/openexchangerates-test-adapter
-
-## 1.0.6
-
-### Patch Changes
-
-- bfa201d6d: Bumped framework version
-- bfa201d6d: Bumped framework version
-- c600ca386: Upgrade typescript version to 5.0.4
-- bfa201d6d: Bumped framework version
-
-## 1.0.5
-
-### Patch Changes
-
-- 9bcb13c90: Bumped framework version
-
-## 1.0.4
-
-### Patch Changes
-
-- 31af84e69: Bumped framework version
-
-## 1.0.3
-
-### Patch Changes
-
-- 77ad946a9: Bumped framework version
-
-## 1.0.2
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- 8fe5be225: Bumped framework version
-
-## 1.0.0
-
-### Major Changes
-
-- 4f16f9a2d: Initial version of the adapter
