--- packages/sources/openexchangerates-test/src/config/includes.json
@@ -1,35 +1,0 @@
-[
-  {
-    "from": "IDR",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "IDR",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "COP",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "COP",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "KRW",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "KRW",
-        "inverse": true
-      }
-    ]
-  }
-]
