--- packages/sources/openexchangerates-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { endpoint as forex } from './forex'
