--- packages/sources/openexchangerates-test/test-payload.json
@@ -1,6 +1,0 @@
-{
- "requests": [{
-  "from": "JPY",
-  "to": "USD"
- }]
-}
