--- packages/sources/openexchangerates/schemas/env.json
@@ -1,22 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/openexchangerates-adapter.json",
-  "title": "Chainlink Open Exchange Rates External Adapter",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "description": "An API key that can be obtained from [here](https://openexchangerates.org/signup)",
-      "required": true
-    },
-    "API_ENDPOINT": {
-      "default": "https://openexchangerates.org/api/",
-      "type": "string"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
