+++ packages/sources/openexchangerates/test/integration/adapter.test.ts
@@ -1,31 +1,40 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import * as process from 'process'
-import { server as startServer } from '../../src'
-import {
-  mockRateResponseSuccess,
-  mockInverseRateResponseSuccess,
-  mockResponseFailure,
-} from './fixtures'
-import { setupExternalAdapterTest } from '@chainlink/ea-test-helpers'
-import type { SuiteContext } from '@chainlink/ea-test-helpers'
 import { SuperTest, Test } from 'supertest'
+import { setupExternalAdapterTest, SuiteContext } from './setup'
+import { ServerInstance } from '@chainlink/external-adapter-framework'
+import { mockInverseRateResponseSuccess, mockRateResponseSuccess } from './fixtures'
 
 describe('execute', () => {
   const id = '1'
+  let spy: jest.SpyInstance
+  beforeAll(async () => {
+    const mockDate = new Date('2022-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+  })
+
+  afterAll((done) => {
+    spy.mockRestore()
+    done()
+  })
+
   const context: SuiteContext = {
     req: null,
-    server: startServer,
+    server: async () => {
+      process.env['RATE_LIMIT_CAPACITY_SECOND'] = '5000'
+      process.env['METRICS_ENABLED'] = 'false'
+      process.env['API_KEY'] = 'fake-api-key'
+      const server = (await import('../../src')).server
+      return server() as Promise<ServerInstance>
+    },
   }
 
   const envVariables = {
     CACHE_ENABLED: 'false',
-    API_KEY: process.env.API_KEY || 'fake-api-key',
   }
 
   setupExternalAdapterTest(envVariables, context)
 
-  describe('forex rate api', () => {
-    const data: AdapterRequest = {
+  describe('forex endpoint', () => {
+    let data = {
       id,
       data: {
         base: 'ETH',
@@ -45,18 +54,15 @@ describe('execute', () => {
         .expect(200)
       expect(response.body).toMatchSnapshot()
     })
-  })
 
-  describe('forex rate api with batched inverse pairs', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        base: 'IDR',
-        quote: ['USD', 'ETH'],
-      },
-    }
-
-    it('should return success', async () => {
+    it('inverse pair should return success', async () => {
+      data = {
+        id,
+        data: {
+          base: 'IDR',
+          quote: 'USD',
+        },
+      }
       mockInverseRateResponseSuccess()
 
       const response = await (context.req as SuperTest<Test>)
@@ -68,19 +74,16 @@ describe('execute', () => {
         .expect(200)
       expect(response.body).toMatchSnapshot()
     })
-  })
 
-  describe('forex rate api with invalid base', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        base: 'NON-EXISTING',
-        quote: 'USD',
-      },
-    }
-
-    it('should return failure', async () => {
-      mockResponseFailure()
+    it('should return error for invalid quote', async () => {
+      mockRateResponseSuccess()
+      data = {
+        id,
+        data: {
+          base: 'ETH',
+          quote: 'INVALID',
+        },
+      }
 
       const response = await (context.req as SuperTest<Test>)
         .post('/')
@@ -88,7 +91,7 @@ describe('execute', () => {
         .set('Accept', '*/*')
         .set('Content-Type', 'application/json')
         .expect('Content-Type', /json/)
-        .expect(200)
+        .expect(502)
       expect(response.body).toMatchSnapshot()
     })
   })
