+++ packages/tsconfig.json
@@ -471,9 +471,6 @@
       "path": "./sources/openexchangerates"
     },
     {
-      "path": "./sources/openexchangerates-test"
-    },
-    {
       "path": "./sources/orchid-bandwidth"
     },
     {
