+++ packages/tsconfig.test.json
@@ -471,9 +471,6 @@
       "path": "./sources/openexchangerates/tsconfig.test.json"
     },
     {
-      "path": "./sources/openexchangerates-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/orchid-bandwidth/tsconfig.test.json"
     },
     {
